edx: Make the script work with arbitrary sites.
OK, not so arbitrary. We still need to know if the course uses newstyle or
oldstyle authentication (e.g., `login` or `login_ajax`). To avoid having
this hardcoded, we can try both alternatives, but this may make the site
owners mad due to hitting their sites constantly.

Caching credentials and using the `requests` library would probably be a
good alternative, from my experience with `coursera-dl`.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>