requirements: Add file with our dependencies.
There are probably more dependencies to include in this file (e.g., include
youtube-dl here).  Also, perhaps we need to update the version of
beautifulsoup4 to something higher.

It is, at least, a starting point for further development.

This is (more or less) one prerequisite to creating a `setup.py` file for
distribution of the code in the PyPI.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>