edx-dl: Split complex list comprehension into functions.
The current list comprehension is doing too many things and it is very
cluttered to read.

In particular, it contains 2 distinct traversals of the list of "soup
markup" returned by BeautifulSoup 4, iteration with an index and creation of
named tuples.

This commit splits functions so that more parts in the action have names.

While we are at it, we also get rid of variables with names that are all
caps.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>