Add initial setup.py file.
This makes the project distributable on PyPI. Note that the file still has a
considerable amount of FIXME's, which mark things that we should decide how
to solve.

One of those is the distribution of the documentation in `.rst` format.  The
easiest way to perform that is to simply use

    pandoc --from=markdown --to=rst --output=README.rst README.md

before creating a package with `python setup.py sdist`.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>