Corrected subsection/unit parsing
The previous versions of the script didn't extract all the subsections
to find the resources. I added a subsections attribute to the sections
tuple and extract all the information in a given step (instead of doing
it multiple times as the script did before).

Aditionally, I created a new dict that will map an url with all the
downloadable resources (units). {url: unit}. This will also help to have
a reference to already parsed resources and can be used to extract units
while iteratin via sections or subsections.