edx_dl: _edx_get_headers: Remove initial underscore from function name.
Without this, we can't import (without mangling the name) the function in
the test suite.

With this change, we can make our tests pass for the first time and hook up
our test suite with a continuous integration system like Travis CI and be
more confident with our refactoring work.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>