Deleted temporary function to transform unit information into urls.
I also refactored a bit the downloade process, it can seem more complex,
but it is not at all, it is really useful to do it like this since we
can now show nicer information to the user (e.g. show which section are
we processing or add correct prefixes to the filenames).