tox: Add configuration for automation of tests.
In my limited experience, tox is a good tool for:

* Creating virtual environments for different versions of python.
* Gathering the dependencies of the project (e.g., bs4).
* Running the unit tests for each virtual environments.

All this with only a single command-line command: `tox`.

You can think of tox (at least in the sense that I described above) as a
"local version" of Travis that you can any time you wish, without having to
push your code to the repository for commits.

You only need to have the Python versions that you want to test installed
for it to work.

Furthermore, with this change, I was able to test that the project works
with pypy, which is super cool, since that's (perhaps) the future.  If I had
a sufficiently new jython version installed here, I would love to test it
also.  (Alas, Debian only has jython 2.5 available).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>