Refactored display courses functions to show only available courses.
I changed the default to available courses only since it does not make
sense to the user to show courses that he can not download at the
moment. Aditionally, the courses who are not available don't have
complete information and can introduce further issues.

I also added the course id as part of the course string, this will allow
the user to identify the course, in case he want to download it by hand
later on.