edx_dl: Simplify custom execution of external commands.
This kills the custom and over-complicated way of executing external
commands with a simple call to Python's `subprocess.call`.

We should check what the return code of the execution is (this is a TODO),
but, at least, we don't eat the complete output fo the external programs in
case of errors (which was happening to this committer, by the way).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>