edx_dl/edx_dl: Refactor function `edx_json2srt`.
We avoid repeated string concatenations (which has quadratic complexity, or
\Omega(n^2)) and create a list with the strings of the subtitle, to return
it in one one (which is O(n)).

While we are at it, we also refactor a "constant", to avoid repeated
creation of objects, since only one `datetime` is enough for the creation of
the other ones.

In my opinion, this is the first candidate function to be part of the test
suite.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>