edx_dl/edx_dl: Fix bug calling youtube-dl with empty parameter.
If we don't want subtitles, then we were passing an empty string to
youtube-dl.  This is wrong and causes failures for the user as the following
messages show:

```
[info] Output directory: Downloaded
ERROR: u'' is not a valid URL. Set --default-search "ytsearch" (or run  youtube-dl "ytsearch:" ) to search YouTube
ERROR: u'' is not a valid URL. Set --default-search "ytsearch" (or run  youtube-dl "ytsearch:" ) to search YouTube
(...)
```

This is, at least, "less bad" now that we call `subprocess.call` directly
instead of using a custom function to call external programs like we did
before (which ended up eating those error messages, as I related in my first
message at https://github.com/shk3/edx-downloader/issues/208).

Hopefully, this fixes all the problems invoking youtube-dl that we know of
so far.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>