edx_dl/{edx_dl,compat}: Rename _print function to be accessible across modules.
This fixes the problems that we have with the explicit need of importing our
overriding of the builtin function.

Since the functions are different, we shouldn't be clobbering the functions,
anyway. We also eliminate some nasty hacks.

The code is probably worse from a PEP-8 perspective, but this should be easy
to deal with.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>