test_utils: Add test for test_get_filename_from_prefix.
Another silly function. Note that its result depends on the argument to it
*not* being a prefix of any other file *and* the order in which the
filesystem returns the results.

For instance, in the current root directory of our project, we have two
files with names `requirements.txt` and `requirements-dev.txt`.  If we only
pass the prefix as `requirements`, the result of the file is unspecified, as
determined by the Python documentation:

    https://docs.python.org/2/library/os.html#os.listdir

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>