test_parsing: Initiate change to pytest.
Note that while this may *seem* noisy, there are very little changes
involved.

To see better what is being done, use `git diff --ignore-all-space` and it
will be clear that most changes are just arrangements of the indentation
(since we don't need the wrapper class anymore) and, also, the removal of
`self` (since it doesn't exist without the test class).

So, the changes are really quite small here. Further patches will migrate to
a more py.test idiomatic style.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>