edx_dl: Catch exception that prevented automated downloads until the end.
We need to decide if we introduce a command-line option to let the user
control if they want to ignore or not the error conditions (youtube-dl does
this).

With this in place, we should also check the return value that youtube-dl
returns or, in general, any external command that we execute.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>