edx_dl: Shorten the body of num_urls_in_units_dict.
This still sucks beyond anything beliefs, but it is shorter and
(potentially) easier to read, while being (hopefully) more readable.

As a side comment, the bytecode generated by this version (with python 2.7)
is smaller than for the cascade of if's. Not that it matters, but it is nice
to know. :)

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>