edx_dl/edx_dl: Give better output when we can't download something.
Instead of silently failing, we should loudly tell the user what the problem
was, since this will help them provide us with useful information so that we
can help them (or they can fix the problem by themselves).

Since downloads get interrupted in case of errors, we should add an option
on the command line so that the user can choose to ignore such errors at his
peril.

This is, BTW, the strategy that youtube-dl adopts with its option named
`--ignore-errors`.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>