edx_dl/edx_dl: Don't fail silently when calling external commands.
In the same spirit as the previous change, we loudly complain when we had
problems executing the external command.

This way, issues like #163 and #246 will be easier to track down.  Again, we
should allow the user to ignore the errors, but this will come later, once
we decide which default should be a sane default.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>