test_utils: Make test_utils pass again.
Since we now raise exceptions when execution of an external command fails,
we split the test in two (which is good) and adapt the part where we expect
something to be raised to handle the raised exception.

Note that I have chosen to *not* use the unittest assertRaises method, since
I plan on changing everything to py.test in the very near future.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>