edx_dl/{edx_dl,utils}: Add option to ignore errors.
This way, the user can choose if they want to ignore some errors that occur
during download of large courses, so that the process actually finishes.

It should be noted that the output of edx-dl is starting to get *too*
cluttered with all those uninteresting messages shown to the user.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>