Convert program to use the logging module instead of compat_print.
Instead of our home-grown function `compat_print`, we use the standard
library module `logging` for sending messages of varying importance to the
users.

This avoids duplication of code of checking which verbosity level we want
and if such and such message should be printed or not, as all that decision
is already built-in by people more experienced than us.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>