edx_dl/edx_dl: Add whitespace for legibility.
We should move the validation of the command line options either to the
`parse_args` function or, if it is going to be too bloated, concentrate the
validation in a separate function.

The way that we have now (with things scattered all over the place) makes it
hard to modularize the code.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>