Add warning to password command line option
Command line options are usually visible to other user on the same
system and so the password could bee seen there. Improper permissions on
the command line history file might be an other source of exposing this
password.

At least warn the user reading the help screen about the risks of using
this option.