Add --export-filename, --export-format options (fix #260)
These options allow to save collected URLs into a file and use in an external
download manager. If --export-filename option is specified, download is not
performed.

Use dash "-" in --export-filename to output URLs to stdout.

--export-format option uses old-style Python formatting rules. The only
supported variable that can be used in the option is "%(url)s". This one
includes both URL from a Video or from a "resource url list".