Do not use reason field of URLError exception (fix #251)
In the code we catch URLError exception. Users, however,
sometimes get HTTPError exception which is a subclass of
URLError. In certain versions of Python, HTTPError does
not have reason field. According to the documentation,
at least version 2.6 lacks it:

https://docs.python.org/2.6/library/urllib2.html#urllib2.HTTPError

A very common and simple solution is to simply format the
whole exception into a string.