edx_dl/common: Use comments on the classes as the module docstring.
The huge comment that we have is better if it is made part of the docstring
of the module, as that is informative when using pydoc (or any other kind of
documentation browser that one wishes to use).

In fact, I would argue that it is even going to be nice to have it become
"part of our API" (for the lack of a better term), as I believe that it will
bring us more flexibility.

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>